requirements.txt
discord.py
pandas
sqlite3
